Fix CI structural limits and docs content schema
